Update version to 1.0.2 Fix problems with finding special commands like : and . for help added a options section in the help file Moved help to internal command (from builtin) Added some error testing for tests Converted unix xmlsh to LF format hope it stays

+++ xmlsh/test/run_test.xsh
@@ -2,15 +2,17 @@
 
 [ $# -ne 1 ] && exit 1
 
+[ -f _out.txt ] && rm _out.txt
+[ -f _err.txt ] && rm _err.txt
 
 
-
-$1  > _out.txt 
+$1  > _out.txt 2> _err.txt
 RET=$?
 if [ $RET -ne 0 ] ; then
    echo $1 failed: return code: $RET
    exit $RET
 fi     
+
 
 if [ -f out/${1}.out ] ; then 
 	xcmp -b _out.txt out/${1}.out 
@@ -18,17 +20,30 @@ if [ -f out/${1}.out ] ; then
 		echo $1 out/${1}.out different output
 		exit 1
 	fi
-	rm _out.txt
+	[ -f _out.txt ] && rm _out.txt
+
 elif [ -f out/${1}.xml ] ; then 
 	xcmp -x -b _out.txt out/${1}.xml 
 	if [ $? -ne 0 ] ; then
 		echo $1 out/${1}.xml different output
 		exit 1
 	fi
-	rm _out.txt
+	[ -f _out.txt ] && rm _out.txt
+
 else
 	echo $1 - no output to compare 
  	exit 2
 fi
+
+if [ -f _err.txt -a -f out/${1}.err ] ; then
+	xcmp -b _err.txt out/${1}.err 
+	if [ $? -ne 0 ] ; then
+		echo $1 out/${1}.err different output
+		exit 1
+	fi
+fi
+	
+[ -f _err.txt ] && rm _err.txt
+
 
 exit 0
